pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pooli-aiops-monitoring"
version = "0.1.0"
description = "PyOD-based AIOps scaffold for pooli-be monitoring"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "joblib>=1.4",
  "numpy>=1.26",
  "pandas>=2.2",
  "pyod>=2.0",
  "PyYAML>=6.0",
  "requests>=2.32",
  "scikit-learn>=1.5",
]

[project.scripts]
pooli-aiops = "pooli_aiops.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
